cleaning out some EOVSA code as we know we dont want to include it in elsa and anna. this is just back end tidying up
